move errors to shared package
