+++ apps/web/src/components/CommunityHolderGrid/CommunityHolderGridItem.tsx
@@ -8,11 +8,11 @@ import InteractiveLink from '~/components/core/InteractiveLink/InteractiveLink';
 import { VStack } from '~/components/core/Spacer/Stack';
 import { BaseM } from '~/components/core/Text/Text';
 import { useModalActions } from '~/contexts/modal/ModalContext';
-import { CouldNotRenderNftError } from '~/errors/CouldNotRenderNftError';
 import { CommunityHolderGridItemFragment$key } from '~/generated/CommunityHolderGridItemFragment.graphql';
 import { CommunityHolderGridItemQueryFragment$key } from '~/generated/CommunityHolderGridItemQueryFragment.graphql';
 import TokenDetailView from '~/scenes/TokenDetailPage/TokenDetailView';
 import { useReportError } from '~/shared/contexts/ErrorReportingContext';
+import { CouldNotRenderNftError } from '~/shared/errors/CouldNotRenderNftError';
 import { getOpenseaExternalUrl } from '~/utils/getOpenseaExternalUrl';
 import getVideoOrImageUrlForNftPreview from '~/utils/graphql/getVideoOrImageUrlForNftPreview';
 import { graphqlTruncateUniversalUsername } from '~/utils/wallet';
