+++ apps/web/src/components/GalleryEditor/CollectionEditor/StagedNftImage.tsx
@@ -2,13 +2,13 @@ import { graphql, useFragment } from 'react-relay';
 import styled from 'styled-components';
 
 import NftPreviewLabel from '~/components/NftPreview/NftPreviewLabel';
-import { CouldNotRenderNftError } from '~/errors/CouldNotRenderNftError';
 import { StagedNftImageFragment$key } from '~/generated/StagedNftImageFragment.graphql';
 import { StagedNftImageImageFragment$key } from '~/generated/StagedNftImageImageFragment.graphql';
 import { StagedNftImageVideoFragment$key } from '~/generated/StagedNftImageVideoFragment.graphql';
 import { useImageFailureCheck } from '~/hooks/useImageFailureCheck';
 import { useThrowOnMediaFailure } from '~/hooks/useNftRetry';
 import { useReportError } from '~/shared/contexts/ErrorReportingContext';
+import { CouldNotRenderNftError } from '~/shared/errors/CouldNotRenderNftError';
 import getVideoOrImageUrlForNftPreview from '~/utils/graphql/getVideoOrImageUrlForNftPreview';
 
 type Props = {
