+++ apps/web/src/components/GalleryEditor/CollectionSidebar/CollectionListItem.tsx
@@ -13,10 +13,10 @@ import { TitleXS } from '~/components/core/Text/Text';
 import { useGalleryEditorContext } from '~/components/GalleryEditor/GalleryEditorContext';
 import { NewTooltip } from '~/components/Tooltip/NewTooltip';
 import { useTooltipHover } from '~/components/Tooltip/useTooltipHover';
-import { ErrorWithSentryMetadata } from '~/errors/ErrorWithSentryMetadata';
 import { CollectionListItemQueryFragment$key } from '~/generated/CollectionListItemQueryFragment.graphql';
 import HideIcon from '~/icons/HideIcon';
 import ShowIcon from '~/icons/ShowIcon';
+import { ErrorWithSentryMetadata } from '~/shared/errors/ErrorWithSentryMetadata';
 import unescape from '~/utils/unescape';
 
 import useMoveCollectionModal from './useMoveCollectionModal';
