+++ apps/web/src/contexts/boundary/ReportingErrorBoundary.tsx
@@ -8,9 +8,9 @@ import {
 } from 'react';
 import { Primitive } from 'relay-runtime/lib/store/RelayStoreTypes';
 
-import { CouldNotRenderNftError } from '~/errors/CouldNotRenderNftError';
-import { ErrorWithSentryMetadata } from '~/errors/ErrorWithSentryMetadata';
 import { ErrorReportingContext } from '~/shared/contexts/ErrorReportingContext';
+import { CouldNotRenderNftError } from '~/shared/errors/CouldNotRenderNftError';
+import { ErrorWithSentryMetadata } from '~/shared/errors/ErrorWithSentryMetadata';
 
 export type ReportingErrorBoundaryFallbackProps = { error: Error };
