+++ packages/shared/src/errors/ErrorWithSentryMetadata.ts
