+++ packages/shared/src/errors/formatError.ts
